get rid of proxy delegate, pass command instead
