+++ fearless/Modules/Wallet/Contacts/ContactViewModel.swift
@@ -4,6 +4,8 @@ import RobinHood
 import SoraFoundation
 
 final class ContactViewModel: ContactsLocalSearchResultProtocol {
+    var command: WalletCommandProtocol?
+
     var cellReuseIdentifier: String { ContactsConstants.contactCellIdentifier }
     var itemHeight: CGFloat { ContactsConstants.contactCellHeight }
 
@@ -12,32 +14,16 @@ final class ContactViewModel: ContactsLocalSearchResultProtocol {
     let accountId: String
     let image: UIImage?
     let name: String
-    let commandFactory: WalletCommandFactoryProtocol
-
-    weak var delegate: ContactViewModelDelegate?
 
     init(firstName: String,
          lastName: String,
          accountId: String,
          image: UIImage?,
-         name: String,
-         delegate: ContactViewModelDelegate?,
-         commandFactory: WalletCommandFactoryProtocol) {
+         name: String) {
         self.firstName = firstName
         self.lastName = lastName
         self.accountId = accountId
         self.image = image
         self.name = name
-        self.delegate = delegate
-        self.commandFactory = commandFactory
     }
-}
-
-extension ContactViewModel: WalletCommandProtocol {
-    var command: WalletCommandProtocol? { self }
-
-    func execute() throws {
-        delegate?.didSelect(contact: self)
-    }
-
 }
